Added new exercise add feature to trainingsplan, no persistence yet

+++ Gym Companion App/Entitites/Exercise.swift
@@ -18,12 +18,12 @@ import SwiftData
     var maxWeight: Int32 = 0
     var minWeight: Int32 = 0
     var name: String
-    var weightStep: Int32 = 0
+    var weightStep: Float = 0
     var category: ExerciseCategory?
     var executions: [ExerciseExecution]?
     @Relationship(inverse: \ExerciseInTraining.exercise) var inTrainings: [ExerciseInTraining]?
     
-    init(isDefault: Bool = true, explanation: String = "", fullName: String = "", iconName: String = "", maxWeight: Int32 = 0, minWeight: Int32 = 0, name: String = "", weightStep: Int32 = 0, category: ExerciseCategory? = nil, executions: [ExerciseExecution]? = nil, inTrainings: [ExerciseInTraining]? = nil) {
+    init(isDefault: Bool = true, explanation: String = "", fullName: String = "", iconName: String = "", maxWeight: Int32 = 0, minWeight: Int32 = 0, name: String = "", weightStep: Float = 0, category: ExerciseCategory? = nil, executions: [ExerciseExecution]? = nil, inTrainings: [ExerciseInTraining]? = nil) {
         self.isDefault = isDefault
         self.explanation = explanation
         self.fullName = fullName
